fix: move impersonation to superadmin, hide org settings for platform admins
- Remove ImpersonateUserButton from /admin/users/[id] detail page
- Remove inline impersonate action from admin UserTable (no longer a
  client component)
- Add ImpersonateUserButton to /superadmin/users/[id] detail page
- Add impersonate action to superadmin UserTable
- Hide Organization settings nav item from AdminSidebar for superusers
